requirements.txt
asgiref==3.8.1
astroid==3.1.0
black==24.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coverage==7.5.1
dill==0.3.8
Django==5.0.6
django-stubs==5.0.0
django-stubs-ext==5.0.0
djangorestframework==3.15.1
djangorestframework-stubs==3.15.0
drf-yasg==1.21.7
filelock==3.14.0
fsspec==2024.3.1
huggingface-hub==0.23.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
mpmath==1.3.0
mypy==1.10.0
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pylint==3.1.0
pytest==8.2.0
pytest-cov==5.0.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.31.0
safetensors==0.4.3
sqlparse==0.5.0
sympy==1.12
tokenizers==0.19.1
tomlkit==0.12.5
torch==2.2.2
tqdm==4.66.4
transformers==4.40.2
types-PyYAML==6.0.12.20240311
types-requests==2.31.0.20240406
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
